+++ Source_Files/memory.h
@@ -28,11 +28,11 @@ class MEMORY {
   void fileReader(const std::string& filename);
   uint32_t readAddress(const size_t& addr) const;
 
-  uint8_t read8(uint32_t address);
-  uint16_t read16(uint32_t address);
+  uint8_t read8(uint16_t address);
+  uint16_t read16(uint16_t address);
 
-  void write8(uint32_t address, uint8_t data);
-  void write16(uint32_t address, uint16_t data);
+  void write8(uint16_t address, uint8_t data);
+  void write16(uint16_t address, uint16_t data);
 };
 
 #endif
